LPS-88934 Split TypedInput component into separate components
